Narrow test helper compilation
Move shared test helpers out of `tests/common` so each integration test
crate only compiles the helpers it uses. Keep `tests/common` limited to
exported macros, and place CLI and wrapping assertions behind narrower
support modules or local fixtures.

This lets `-D warnings` catch unused test helpers directly instead of
relying on broad lint suppressions.
